current_requirements.txt
absl-py==2.3.1
addict==2.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
crcmod==1.7
cryptography==45.0.7
cycler==0.12.1
filelock==3.14.0
fonttools==4.59.2
fsspec==2025.7.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.74.0
hf-xet==1.1.9
huggingface-hub==0.34.4
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jmespath==0.10.0
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy_loader==0.4
Markdown==3.8.2
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mmcls==0.25.0
mmcv-full==1.7.2
mmengine==0.10.7
mmsegmentation==0.30.0
model-index==0.1.11
mpmath==1.3.0
narwhals==2.3.0
networkx==3.4.2
numpy==1.26.4
opencv-python==4.12.0.88
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.2
pandas==2.3.2
pillow==11.3.0
platformdirs==4.4.0
prettytable==3.16.0
protobuf==6.32.0
pyarrow==21.0.0
pycparser==2.22
pycryptodome==3.23.0
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2023.4
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
requests==2.28.2
rich==13.4.2
rpds-py==0.27.1
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.49.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
tifffile==2025.5.10
timm==1.0.19
toml==0.10.2
tomli==2.2.1
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado==6.5.2
tqdm==4.65.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
yapf==0.43.0
